Implement clients, logs and other useful stuff

+++ hw/server-bench/src/main/java/ru/spbau/bocharov/serverbench/server/impl/udp/FixedSizeThreadPoolUDPServer.java
@@ -11,7 +11,7 @@ public class FixedSizeThreadPoolUDPServer extends BaseUDPServer {
 
     private final ExecutorService pool = Executors.newFixedThreadPool(THREAD_COUNT);
 
-    public FixedSizeThreadPoolUDPServer(short port) {
+    public FixedSizeThreadPoolUDPServer(int port) {
         super(port);
     }
